server/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.7
django-cors-headers==3.4.0
django-grappelli==2.14.2
djangorestframework==3.11.0
djangorestframework-gis==0.15
idna==2.10
itsdangerous==1.1.0
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
